+++ src/index.ts
@@ -1,4 +1,5 @@
 import path from 'path'
+import fs from 'fs'
 import chokidar from 'chokidar'
 import { Configuration as WebpackConfig, Entry as WebpackEntry } from 'webpack'
 // @ts-ignore
@@ -6,45 +7,76 @@ import RuleSet from 'webpack/lib/RuleSet'
 import { Module } from '@nuxt/types'
 
 import { requireNuxtVersion } from './compatibility'
-import { scanComponents } from './scan'
+import { scanComponents, ScanDir } from './scan'
+
+type componentsDirHook = (dirs: ComponentsDir[]) => void | Promise<void>
+type componentsExtendHook = (components: (ComponentsDir|ScanDir)[]) => void | Promise<void>
+
+declare module '@nuxt/types/config/hooks' {
+  interface NuxtConfigurationHooks {
+    'components:dirs'?: componentsDirHook
+    'components:extend'?: componentsExtendHook
+    components?: {
+      dirs?: componentsDirHook
+      extend?: componentsExtendHook
+    }
+  }
+}
+
+export interface ComponentsDir extends ScanDir {
+  watch?: boolean
+  transpile?: 'auto' | boolean
+}
 
 export interface Options {
-  dirs: Array<string | {
-    path: string
-    pattern?: string
-    ignore?: string[]
-    prefix?: string
-    watch?: boolean
-    transpile?: boolean
-  }>
+  dirs: (string | ComponentsDir)[]
 }
 
 const isPureObjectOrString = (val: any) => (!Array.isArray(val) && typeof val === 'object') || typeof val === 'string'
+const getDir = (p: string) => fs.statSync(p).isDirectory() ? p : path.dirname(p)
 
-export default <Module<Options>> function (moduleOptions) {
+export default <Module> function () {
   requireNuxtVersion.call(this, '2.10')
 
   const options: Options = {
+    // @ts-ignore This is expected as default dirs will be overriden by user config
     dirs: ['~/components'],
-    ...moduleOptions,
     ...this.options.components
   }
 
   this.nuxt.hook('build:before', async (builder: any) => {
     const nuxtIgnorePatterns: string[] = builder.ignore.ignore ? builder.ignore.ignore._rules.map((rule: any) => rule.pattern) : /* istanbul ignore next */ []
+
+    await this.nuxt.callHook('components:dirs', options.dirs)
+
     const componentDirs = options.dirs.filter(isPureObjectOrString).map((dir) => {
-      const dirOptions = typeof dir === 'object' ? dir : { path: dir }
+      const dirOptions: ComponentsDir = typeof dir === 'object' ? dir : { path: dir }
+
+      let dirPath = dirOptions.path
+      try { dirPath = getDir(this.nuxt.resolver.resolvePath(dirOptions.path)) } catch (err) { }
+
+      const transpile = typeof dirOptions.transpile === 'boolean' ? dirOptions.transpile : 'auto'
+
+      const enabled = fs.existsSync(dirPath)
+      if (!enabled && dirOptions.path !== '~/components') {
+        // eslint-disable-next-line no-console
+        console.warn('Components directory not found: `' + dirPath + '`')
+      }
+
       return {
         ...dirOptions,
-        path: this.nuxt.resolver.resolvePath(dirOptions.path),
+        enabled,
+        path: dirPath,
         pattern: dirOptions.pattern || `**/*.{${builder.supportedExtensions.join(',')}}`,
-        ignore: nuxtIgnorePatterns.concat(dirOptions.ignore || [])
+        ignore: nuxtIgnorePatterns.concat(dirOptions.ignore || []),
+        transpile: (transpile === 'auto' ? dirPath.includes('node_modules') : transpile)
       }
-    })
+    }).filter(d => d.enabled)
 
     this.options.build!.transpile!.push(...componentDirs.filter(dir => dir.transpile).map(dir => dir.path))
 
-    let components = await scanComponents(componentDirs)
+    let components = await scanComponents(componentDirs, this.options.srcDir!)
+    await this.nuxt.callHook('components:extend', components)
 
     this.extendBuild((config) => {
       const { rules }: any = new RuleSet(config.module!.rules)
@@ -68,13 +100,31 @@ export default <Module<Options>> function (moduleOptions) {
           return
         }
 
-        components = await scanComponents(componentDirs)
+        components = await scanComponents(componentDirs, this.options.srcDir!)
+        await this.nuxt.callHook('components:extend', components)
+
         await builder.generateRoutesAndFiles()
       })
 
       // Close watcher on nuxt close
       this.nuxt.hook('close', () => {
         watcher.close()
+      })
+    }
+
+    // Add templates
+    const templates = [
+      'components/components.js',
+      'components/components.json',
+      'vetur/tags.json'
+    ]
+    for (const t of templates) {
+      this.addTemplate({
+        src: path.resolve(__dirname, '../templates', t),
+        fileName: t,
+        options: {
+          getComponents: () => components
+        }
       })
     }
   })