+++ test/fixture/my-lib/components/index.js
@@ -1,0 +1,1 @@
+throw new Error('no export from index!')