+++ test/fixture/pages/pug.vue
@@ -1,0 +1,10 @@
+<template lang="pug">
+  div
+    p pug loaded components:
+    foo
+    lazy-bar
+    base-button
+    icon-home
+    m-awesome
+
+</template>